changed size return value to from int to long

+++ src/main/java/com/hivemq/spi/services/RetainedMessageStoreCluster.java
@@ -55,6 +55,6 @@ public interface RetainedMessageStoreCluster {
     /**
      * @return the number of all retained messages
      */
-    ListenableFuture<Integer> size();
+    ListenableFuture<Long> size();
 
 }
